REFACTORING models and api controller

+++ BookkeepingNasheDetstvo.Server/Controllers/ApiController.cs
@@ -1,5 +1,6 @@
 using System;
 using System.Collections.Generic;
+using System.Diagnostics.CodeAnalysis;
 using System.Linq;
 using System.Threading.Tasks;
 using BookkeepingNasheDetstvo.Server.Attributes;
@@ -16,6 +17,7 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
 {
     [Route("api")]
     [Produces("application/json")]
+    [SuppressMessage("ReSharper", "UseDeconstructionOnParameter")]
     public class ApiController : Controller
     {
         private readonly BookkeepingContext _context;
@@ -30,7 +32,7 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
             if (teacher == default)
                 return NotFound();
 
-            var credential = await _context.Credentials.Find(c => c.Id == teacher.Id).SingleAsync();
+            var credential = await _context.Credentials.Find(c => c.TeacherId == teacher.Id).SingleAsync();
 
             if (credential.PasswordHash != PasswordExtensions.HashPassword(model.Password, credential.Salt))
                 return StatusCode(403);
@@ -55,17 +57,24 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
 
         [HttpGet("current")]
         [ValidateAccessToken]
-        public ActionResult<Teacher> Current(Teacher current) => current;
+        public Teacher Current(Teacher current)
+        {
+            return current;
+        }
 
         [HttpGet("teachers")]
         [ValidateAccessToken]
-        public Task<List<Teacher>> ListTeachers() =>
-            _context.Teachers.Find(FilterDefinition<Teacher>.Empty).ToListAsync();
+        public Task<List<Teacher>> ListTeachers()
+        {
+            return _context.Teachers.Find(FilterDefinition<Teacher>.Empty).ToListAsync();
+        }
 
         [HttpGet("teacher/{id:required}")]
         [ValidateAccessToken]
-        public Task<Teacher> GetTeacher(string id) =>
-            _context.Teachers.Find(t => t.Id == id).SingleOrDefaultAsync();
+        public Task<Teacher> GetTeacher(string id)
+        {
+            return _context.Teachers.Find(t => t.Id == id).SingleOrDefaultAsync();
+        }
 
         [HttpPost("teacher")]
         [ValidateAccessToken]
@@ -83,6 +92,21 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
             await _context.Teachers.ReplaceOneAsync(t => t.Id == teacher.Id, teacher,
                 new UpdateOptions {IsUpsert = true});
             
+            var newSalt = PasswordGenerator.Generate(8);
+            try
+            {
+                await _context.Credentials.InsertOneAsync(new Credential
+                {
+                    Salt = newSalt,
+                    TeacherId = teacher.Id,
+                    PasswordHash = PasswordExtensions.HashPassword("1", newSalt)
+                });
+            }
+            catch (MongoWriteException)
+            {
+                // ignored
+            }
+
             return teacher.Id;
         }
 
@@ -99,7 +123,7 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
                 return StatusCode(403);
 
             var newSalt = PasswordGenerator.Generate(8);
-            await _context.Credentials.UpdateOneAsync(c => c.Id == model.TeacherId,
+            await _context.Credentials.UpdateOneAsync(c => c.TeacherId == model.TeacherId,
                 Builders<Credential>.Update.Set(c => c.Salt, newSalt)
                     .Set(c => c.PasswordHash, PasswordExtensions.HashPassword(model.NewPassword, newSalt)),
                 new UpdateOptions {IsUpsert = false});
@@ -109,11 +133,17 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
 
         [HttpGet("children")]
         [ValidateAccessToken]
-        public Task<List<Child>> ListChildren() => _context.Children.Find(FilterDefinition<Child>.Empty).ToListAsync();
+        public Task<List<Child>> ListChildren()
+        {
+            return _context.Children.Find(FilterDefinition<Child>.Empty).ToListAsync();
+        }
 
         [HttpGet("child/{id:required}")]
         [ValidateAccessToken]
-        public Task<Child> GetChild(string id) => _context.Children.Find(c => c.Id == id).SingleOrDefaultAsync();
+        public Task<Child> GetChild(string id)
+        {
+            return _context.Children.Find(c => c.Id == id).SingleOrDefaultAsync();
+        }
 
         [HttpPost("child")]
         [ValidateAccessToken]
@@ -133,11 +163,14 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
 
         [HttpPost("child/delete/{id:required}")]
         [ValidateAccessToken]
-        public async Task<ActionResult> DeleteChild(string id)
+        public async Task<ActionResult> DeleteChild(string id, Teacher current)
         {
+            if (!current.EditChildren && !current.IsOwner)
+                return StatusCode(403);
+            
             await _context.Children.DeleteOneAsync(c => c.Id == id);
             await _context.Subjects.UpdateManyAsync(FilterDefinition<Subject>.Empty,
-                Builders<Subject>.Update.PullFilter(s => s.Children, c => c.Id == id));
+                Builders<Subject>.Update.PullFilter(s => s.ChildrenIds, childId => childId == id));
             
             return Ok();
         }
@@ -155,113 +188,69 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
                 return StatusCode(403);
             
             await _context.Teachers.DeleteOneAsync(c => c.Id == id);
-            await _context.Subjects.DeleteManyAsync(s => s.Owner.Id == id);
+            await _context.Subjects.DeleteManyAsync(s => s.OwnerId == id);
             await _context.Sessions.DeleteManyAsync(s => s.OwnerId == id);
-            await _context.Credentials.DeleteOneAsync(c => c.Id == id);
+            await _context.Credentials.DeleteOneAsync(c => c.TeacherId == id);
             
             return Ok();
         }
 
         [HttpGet("subjects")]
         [ValidateAccessToken]
-        public async Task<ActionResult<List<Subject>>> ListSubjects([FromQuery] string date)
+        public Task<List<Subject>> ListSubjects([FromQuery] string date)
         {
-            return await _context.Subjects.Find(s => s.Date == date).ToListAsync();
+            return _context.Subjects.Find(s => s.Date == date).ToListAsync();
         }
 
-        [HttpGet("subjects/teachers")]
+        [HttpGet("subjects/allTeachers")]
         [ValidateAccessToken]
         public async Task<object> ListTeachersForSubjects()
         {
             var teachers = await _context.Teachers.Find(FilterDefinition<Teacher>.Empty)
                 .Project<Teacher>("{FirstName:1, LastName:1}").ToListAsync();
 
-            return teachers.Select(t =>
-            {
-                var name = $"{t.LastName} {(t.FirstName.Length == 0 ? string.Empty : $"{t.FirstName[0]}.")}"
-                    .TrimEnd(' ');
-                return new { t.Id, name };
-            });
+            return teachers.Select(t => new { t.Id, t.FirstName, t.LastName });
         }
 
-        [HttpGet("subjects/children")]
+        [HttpGet("subjects/allChildren")]
         [ValidateAccessToken]
         public async Task<object> ListChildrenForSubjects()
         {
             var children = await _context.Children.Find(FilterDefinition<Child>.Empty)
-                .Project<Child>("{FirstName:1, LastName:1}").ToListAsync();
+                .Project<Child>("{FirstName:1, LastName:1, ImageUrl:1}").ToListAsync();
 
-            return children.Select(c =>
-            {
-                var name = $"{c.LastName} {(c.FirstName.Length == 0 ? string.Empty : $"{c.FirstName[0]}.")}"
-                    .TrimEnd(' ');
-                return new { c.Id, name };
-            });
+            return children.Select(c => new { c.Id, Name = $"{c.LastName} {c.FirstName}".Trim(), c.ImageUrl });
         }
 
         [HttpPost("subject/removeChild")]
         [ValidateAccessToken]
         [ValidateModel]
-        public async Task<ActionResult> RemoveChildFromSubject([FromBody] RemoveSubjectChildModel model, Teacher current)
+        public async Task<ActionResult> RemoveChildFromSubject([FromBody] SubjectChildActionModel model, Teacher current)
         {
             if (!current.EditSubjects && !current.IsOwner)
                 return StatusCode(403);
-            
-            var updatedChildren = await _context.Subjects.FindOneAndUpdateAsync(s => s.Id == model.Id,
-                Builders<Subject>.Update.PullFilter(s => s.Children, c => c.Id == model.ChildId),
-                new FindOneAndUpdateOptions<Subject, List<IdNamePair>>
-                {
-                    Projection = new FindExpressionProjectionDefinition<Subject, List<IdNamePair>>(s => s.Children),
-                    ReturnDocument = ReturnDocument.After
-                });
-            
-            if (updatedChildren.Count == 0)
-                await _context.Subjects.DeleteOneAsync(s => s.Id == model.Id);
+
+            await _context.Subjects.UpdateOneAsync(s => s.Id == model.SubjectId,
+                Builders<Subject>.Update.PullFilter(s => s.ChildrenIds, childId => childId == model.ChildId),
+                new UpdateOptions {IsUpsert = false});
             
             return Ok();
         }
 
-        [HttpPost("subject/addChild")]
-        [ValidateAccessToken]
-        [ValidateModel]
-        public async Task<ActionResult<string>> AddChildToSubject([FromBody] AddSubjectChildModel model, Teacher current)
+        [HttpPost("subject/addChild"), ValidateAccessToken, ValidateModel]
+        public async Task<ActionResult<string>> AddChildToSubject([FromBody] SubjectChildActionModel actionModel, Teacher current)
         {
             if (!current.EditSubjects && !current.IsOwner)
                 return StatusCode(403);
             
-            await _context.Subjects.UpdateOneAsync(
-                s => s.Id == model.Id,
-                Builders<Subject>.Update.Push(s => s.Children, model.Child),
+            await _context.Subjects.UpdateOneAsync(s => s.Id == actionModel.SubjectId,
+                Builders<Subject>.Update.AddToSet(s => s.ChildrenIds, actionModel.ChildId),
                 new UpdateOptions {IsUpsert = false});
 
             return Ok();
         }
 
-        [HttpPost("subject/create")]
-        [ValidateAccessToken]
-        [ValidateModel]
-        public async Task<ActionResult<string>> CreateSubject([FromBody] CreateSubjectModel model, Teacher current)
-        {
-            if (!current.EditSubjects && !current.IsOwner)
-                return StatusCode(403);
-
-            var id = ObjectId.GenerateNewId().ToString();
-            await _context.Subjects.InsertOneAsync(new Subject
-            {
-                Id = id,
-                Date = model.Date,
-                Time = model.Time,
-                Owner = model.Owner,
-                IsConsultation = model.IsConsultation,
-                PlaceIdentifier = model.PlaceIdentifier,
-                Children = new List<IdNamePair>()
-            });
-            
-            return id;
-        }
-
-        [HttpGet("statistic/child/{childId:required}")]
-        [ValidateAccessToken]
+        [HttpGet("statistic/child/{childId:required}"), ValidateAccessToken]
         public async Task<ActionResult<object>> GetChildStatistic(DateRangeModel model, string childId, Teacher current)
         {
             if (!current.ReadGlobalStatistic && !current.IsOwner)
@@ -272,13 +261,13 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
                 return NotFound();
 
             var subjects = await _context.Subjects.Find(FilterDefinition<Subject>.Empty)
-                .Project<Subject>("{Children:1,Owner:1,Date:1,_id:0}").ToListAsync();
-            var totalHoursByTeachers = new Dictionary<string, (string name, int total)>();
-            var totalHoursGlobal = 0;
+                .Project<Subject>("{Children:1,OwnerId:1,Date:1,_id:0}").ToListAsync();
+            var totalHoursByTeachers = new Dictionary<string, (int Private, int Group, int Consultation)>();
+            int totalHoursPrivate = 0, totalHoursGroup = 0, totalHoursConsultation = 0;
             
             foreach (var subject in subjects)
             {
-                var childInSubject = subject.Children.Find(c => c.Id == child.Id);
+                var childInSubject = subject.ChildrenIds.Find(cId => cId == child.Id);
                 if (childInSubject == default)
                     continue;
 
@@ -286,32 +275,54 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
                     subjectDate < model.From ||
                     subjectDate > model.To) continue;
                 
-                if (totalHoursByTeachers.TryGetValue(subject.Owner.Id, out var value))
-                    totalHoursByTeachers[subject.Owner.Id] = (value.name, value.total + 1);
-                else totalHoursByTeachers.Add(subject.Owner.Id, (subject.Owner.Name, 1));
+                if (subject.IsConsultation)
+                    totalHoursConsultation++;
+                else if (subject.ChildrenIds.Count == 1)
+                    totalHoursPrivate++;
+                else totalHoursGroup++;
                 
-                ++totalHoursGlobal;
+                totalHoursByTeachers.TryGetValue(subject.OwnerId, out var value);
+                if (subject.IsConsultation)
+                    value.Consultation++;
+                else if (subject.ChildrenIds.Count == 1)
+                    value.Private++;
+                else value.Group++;
+                totalHoursByTeachers[subject.OwnerId] = value;
             }
             
             var teachers = await _context.Teachers.Find(FilterDefinition<Teacher>.Empty)
-                .Project<Teacher>("{ImageUrl:1}").ToListAsync();
+                .Project<Teacher>("{ImageUrl:1,LastName:1,FirstName:1}").ToListAsync();
             var statistic = new
             {
                 child.PerHour, child.PerHourGroup,
-                sourceItems = totalHoursByTeachers.Select(pair =>
+                Items = totalHoursByTeachers.Select(pair =>
                 {
-                    var (id, (name, totalHours)) = pair;
-                    var imageUrl = teachers.Find(t => t.Id == id)?.ImageUrl ?? string.Empty;
-                    return new { totalHours, id, name, imageUrl };
+                    var teacher = teachers.Find(t => t.Id == pair.Key);
+                    return new
+                    {
+                        TotalHours = new
+                        {
+                            pair.Value.Private,
+                            pair.Value.Group,
+                            pair.Value.Consultation
+                        },
+                        Id = pair.Key,
+                        Name = $"{teacher?.LastName} {teacher?.FirstName}".Trim(),
+                        ImageUrl = teacher?.ImageUrl ?? string.Empty
+                    };
                 }),
-                totalHours = totalHoursGlobal,
-                name = $"{child.LastName} {child.FirstName}".Trim()
+                TotalHours = new
+                {
+                    Private = totalHoursPrivate,
+                    Group = totalHoursGroup,
+                    Consultation = totalHoursConsultation
+                },
+                Name = $"{child.LastName} {child.FirstName}".Trim()
             };
             return statistic;
         }
         
-        [HttpGet("statistic/teacher/{teacherId:required}")]
-        [ValidateAccessToken]
+        [HttpGet("statistic/teacher/{teacherId:required}"), ValidateAccessToken]
         public async Task<ActionResult<object>> GetTeacherStatistic(DateRangeModel model, string teacherId, Teacher current)
         {
             if (current.Id != teacherId && !current.ReadGlobalStatistic && !current.IsOwner)
@@ -323,38 +334,62 @@ namespace BookkeepingNasheDetstvo.Server.Controllers
             if (teacher == default)
                 return NotFound();
 
-            var subjects = await _context.Subjects.Find(s => s.Owner.Id == teacher.Id)
-                .Project<Subject>("{Date:1,Children:1,_id:0}").ToListAsync();
-            var totalHoursByChildren = new Dictionary<string, (string name, int total)>();
-            var totalHoursGlobal = 0;
+            var subjects = await _context.Subjects.Find(s => s.OwnerId == teacher.Id)
+                .Project<Subject>("{Date:1,Children:1,IsConsultation:1,_id:0}").ToListAsync();
+            var totalHoursByChildren = new Dictionary<string, (int Private, int Group, int Consultation)>();
+            int totalHoursPrivate = 0, totalHoursGroup = 0, totalHoursConsultation = 0;
             
             foreach (var subject in subjects)
             {
                 var subjectDate = DateTime.Parse(subject.Date);
                 if (subjectDate < model.From || subjectDate > model.To) continue;
+
+                if (subject.IsConsultation)
+                    totalHoursConsultation++;
+                else if (subject.ChildrenIds.Count == 1)
+                    totalHoursPrivate++;
+                else totalHoursGroup++;
                     
-                foreach (var child in subject.Children)
+                foreach (var childId in subject.ChildrenIds)
                 {
-                    if (totalHoursByChildren.TryGetValue(child.Id, out var value))
-                        totalHoursByChildren[child.Id] = (value.name, value.total + 1);
-                    else totalHoursByChildren.Add(child.Id, (child.Name, 1));
-                    
-                    ++totalHoursGlobal;
+                    totalHoursByChildren.TryGetValue(childId, out var value);
+                    if (subject.IsConsultation)
+                        value.Consultation++;
+                    else if (subject.ChildrenIds.Count == 1)
+                        value.Private++;
+                    else value.Group++;
+                    totalHoursByChildren[childId] = value;
                 }
             }
 
-            var children = await _context.Children.Find(FilterDefinition<Child>.Empty).Project<Child>("{ImageUrl:1}").ToListAsync();
+            var children = await _context.Children.Find(FilterDefinition<Child>.Empty)
+                .Project<Child>("{ImageUrl:1,FirstName:1,LastName:1}").ToListAsync();
             return new
             {
                 teacher.PerHour, teacher.PerHourGroup,
-                sourceItems = totalHoursByChildren.Select(pair =>
+                Items = totalHoursByChildren.Select(pair =>
                 {
-                    var (id, (name, totalHours)) = pair;
-                    var imageUrl = children.Find(c => c.Id == id)?.ImageUrl ?? string.Empty;
-                    return new { totalHours, id, name, imageUrl };
+                    var child = children.Find(c => c.Id == pair.Key);
+                    return new
+                    {
+                        Id = pair.Key,
+                        Name = $"{child?.LastName} {child?.FirstName}".Trim(),
+                        ImageUrl = child?.ImageUrl ?? string.Empty,
+                        TotalHours = new
+                        {
+                            pair.Value.Private,
+                            pair.Value.Group,
+                            pair.Value.Consultation
+                        }
+                    };
                 }),
-                totalHours = totalHoursGlobal,
-                name = $"{teacher.LastName} {teacher.FirstName}".Trim()
+                TotalHours = new
+                {
+                    Private = totalHoursPrivate,
+                    Group = totalHoursGroup,
+                    Consultation = totalHoursConsultation
+                },
+                Name = $"{teacher.LastName} {teacher.FirstName}".Trim()
             };
         }
     }
